Clean .bgw
